Agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "chromadb>=1.0.15",
    "fastapi>=0.115.14",
    "langchain-core>=0.3.68",
    "langchain-google-genai>=2.1.6",
    "langgraph>=0.5.1",
    "pandas>=2.3.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.0.0",
    "uvicorn[standard]>=0.35.0",
]
